The return of the Video Embed Code field
A few lines of code to support the Video Embed Code custom field entry and display had been removed going from version 2.x to 3.x. This impacts those who used the Video Embed Code field in the past when upgrading and prevents them from upgrading to latest version which is not ideal.