The Return of the Video Embed Code
A few lines of code to support the Video Embed Code custom field entry and display had been removed going from version 2.x to 3.x. This impacts those who relied upon the Video Embed Code field in the past when upgrading and prevents them from upgrading to latest version which is not ideal.